# perf: micro benchmark worker pool in prices