This shows that worker pool limitation is not the cause of performance regression by benchmarking `GetPrices` query with all prices.

The difference between 10 and 100 workers for CPU time is not statistically significant. At the same time, 400 workers showed an 8% regression relative to 10:
```go
benchstat old_bench new_bench_400
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/tokens/usecase
cpu: Intel(R) Xeon(R) Platinum 8358 CPU @ 2.60GHz
            │  old_bench   │           new_bench_400            │
            │    sec/op    │   sec/op     vs base               │
GetPrices-4   80.46µ ± 14%   87.00µ ± 6%  +8.13% (p=0.035 n=10)

            │  old_bench   │            new_bench_400            │
            │     B/op     │     B/op      vs base               │
GetPrices-4   15.75Ki ± 0%   16.88Ki ± 0%  +7.17% (p=0.000 n=10)

            │ old_bench  │           new_bench_400           │
            │ allocs/op  │ allocs/op   vs base             
```